flipping layer order

+++ src/oval.rs
@@ -5,7 +5,7 @@ pub struct Oval {
     pub pos: (f32, f32),
     pub rotation: f32,
     pub color: (f32, f32, f32, f32),
-    pub layer: u32,
+    pub layer: u16,
     pub detail: u32,
 }
 
@@ -35,7 +35,7 @@ impl Oval {
         self
     }
 
-    pub fn layer(&mut self, layer: u32) -> &mut Self {
+    pub fn layer(&mut self, layer: u16) -> &mut Self {
         self.layer = layer;
         self
     }
